pyproject.toml
[project]
name = "resume_generator"
version = "0.1.0"
description = "resume-generator using crewAI"
authors = [{ name = "Hugo Alejandro Apaza Huaycho", email = "apazahuaychohugoalejandro@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "boto3>=1.38.27",
    "crewai[tools]>=0.105.0,<1.0.0",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "starlette>=0.46.2",
    "uuid>=1.30",
    "uvicorn[standard]>=0.34.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N"]
ignore = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"


[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.8",
]

[tool.hatch.build.targets.wheel]
include = ["src"]
